Demo Materials
This package is the configuration, app and supporting materials for
using closed caption text to explore video.

Materials included are as follows:

package_2015-7-30_15-45-28.zip
the server configuration for the databases and apps

App Configuration.pdf and Database Configuration.pdf
the settings of the http server and the settings of the content
database for reference

ProdSubs
the on-screen text with timestamps that allow for video search

App
the application itself that is to be published with roxy against the
configured http port (currently 8092)

Paul Demo Cert.xml
the workspace for loading ProdSubs. Use the “Get Subtitles for Apache
Web Server” snippet in the workspace to map to the location of the
ProdSubs directory to load all the search documents